+++ wasppy/test/test_swig_bindings.py
@@ -1,59 +1,167 @@
 import unittest
 from wasp import *
+from Database import InputObject, storeFloat, storeStr, ExistsConstraint
 import math
 
 class LinearModel:
+    Definition = None
+    @staticmethod
+    def definition():
+        if LinearModel.Definition is not None: return LinearModel.Definition
+        dens = InputObject(Desc="Salt density")
+        dens.createRequiredSingle("A", Desc="Density A Coefficient").createRequiredSingle("value", Action=storeFloat)
+        dens.createRequiredSingle("B", Desc="Density B Coefficient").createRequiredSingle("value", Action=storeFloat)
+        dens.createSingle("C", Default=1.0, Desc="Density C Coefficient").createRequiredSingle("value", Action=storeFloat)
+        dens.createRequiredSingle("MinTemp", Desc="Minimum temperature").createRequiredSingle("value", MinValExc=0, Action=storeFloat)
+        # TODO add max temperature MinValExc dependency support
+        dens.createRequiredSingle("MaxTemp", Desc="Maximum temperature").createRequiredSingle("value", Action=storeFloat)
+        dens.createSingle("Type", Default="linear", Desc="interpolation type").createRequiredSingle("value", Enums=["linear"], Action=storeStr)
+
+        LinearModel.Definition = dens
+        return LinearModel.Definition
+
+    @staticmethod
+    def createFrom(do:'DeserializedObject'):
+
+        result = LinearModel()
+        result._a = do["A"].value()
+        result._b = do["B"].value()
+        result._c = do["C"].value()
+        result._minT = do["MinTemp"].value()
+        result._maxT = do["MaxTemp"].value()
+
+        theType = do["Type"].value()
+
+        # Require Type to be a supported enumeration
+        # This demonstrates post deserialization diagnostic generation
+        enumerations = result.definition()["Type"]["value"].enumerations()
+        if theType not in enumerations:
+            do.interpreter.createWarningDiagnostic(do["Type"].node, "has value of "+str(do["Type"].node)+" which is not listed in "+str(enumerations))
+
+        return result
+
     ''' _b*x + _c*y = _a '''
-    def __init__(self,params):
+    def __init__(self,params=None):
         self._a = 0.0
         self._b = 1.0
         self._c = 1.0
         self._minT = math.inf
         self._maxT = -math.inf
 
-        for it in params:
-            if it.name() == "MinTemp":
-                self._maxT = float(it)
-            elif it.name() == "MaxTemp":
-                self._minT = float(it)
-            elif it.name() == "A":
-                self._a = float(it)
-            elif it.name() == "B":
-                self._b = float(it)
-            elif it.name() == "C":
-                self._c = float(it)
+        if params is not None:
+            for it in params:
+                if it.name() == "MinTemp":
+                    self._maxT = float(it)
+                elif it.name() == "MaxTemp":
+                    self._minT = float(it)
+                elif it.name() == "A":
+                    self._a = float(it)
+                elif it.name() == "B":
+                    self._b = float(it)
+                elif it.name() == "C":
+                    self._c = float(it)
 
     def get_y(self,x: float) -> "float":
         if self._c == 0.0:
             return math.inf
 
         return (self._a - (self._b * x)) / self._c
+    def __str__(self):
+        return "density - a:"+str(self._a)+" b:"+str(self._b)+" c:"+str(self._c)+" min:"+str(self._minT)+" max:"+str(self._maxT)
 
 class Salt:
-    def __init__(self,params):
+    Definition = None
+    @staticmethod
+    def definition():
+        '''
+            return inputObject - the definition of this object
+
+        '''
+        if Salt.Definition is not None: return Salt.Definition
+        salt = InputObject(Desc="Single Salt instance")
+        salt.createRequiredSingle("id", Enums=["LiF", "NaF", "CaF2", "NH4F", "NaCl"], Desc="Salt type", Action=storeStr)
+        salt.createRequiredSingle("BoilTemp", Desc="Salting boiling temperature") \
+                .createRequiredSingle("value", MinValExc=0, Action=storeFloat)
+        salt.createRequiredSingle("MeltTemp", Desc="Salt melting temperature").createRequiredSingle("value", Action=storeFloat)
+        salt.createSingle("MolecularWeight", Desc="Salt's molecular weight").createRequiredSingle("value", MinValExc=0, Action=storeFloat)
+        salt.addRequiredSingle("Density", LinearModel.definition())
+
+        Salt.Definition = salt
+        return Salt.Definition
+
+    @staticmethod
+    def createFrom(do:'DeserializedObject'):
+        '''
+            deserializedObject - Salt object data deserialized from user input
+            Create a Salt object from the given data and return it to the caller
+        '''
+
+        result = Salt()
+        result._name = do["id"] # not an id=value, just salt(id)
+        result._molew = do["MolecularWeight"].value() # is a key=value MolecularWeight=value
+        result._meltT = do["MeltTemp"].value()
+        result._boilT = do["BoilTemp"].value()
+        result._density = LinearModel.createFrom(do["Density"])
+
+        return result
+
+    def __init__(self,params=None):
         self._name = ""
         self._molew = 0.0
         self._meltT = 0.0
         self._boilT = 0.0
         self._density: LinearModel
 
-        #Ordering of params is only guaranteed if the input order is enforced
-        #For this reason an iterator is used. Should order be guaranteed
-        #attributes could be assigned by node index, i.e. self._name = str(params[i])
-        for it in params:
-            if it.name() == "id":
-                self._name = str(it)
-            elif it.name() == "MolecularWeight":
-                self._molew = float(it)
-            elif it.name() == "MeltTemp":
-                self._meltT = float(it)
-            elif it.name() == "BoilTemp":
-                self._boilT = float(it)
-            elif it.name() == "Density":
-                self._density = LinearModel(it)
+        # Process parse tree parameters if provided
+        if params is not None:
+            for it in params:
+                if it.name() == "id":
+                    self._name = str(it)
+                elif it.name() == "MolecularWeight":
+                    self._molew = float(it)
+                elif it.name() == "MeltTemp":
+                    self._meltT = float(it)
+                elif it.name() == "BoilTemp":
+                    self._boilT = float(it)
+                elif it.name() == "Density":
+                    self._density = LinearModel(it)
+
 
     def density(self,T: float) -> "float":
         return self._density.get_y(T)
+
+    def __str__(self):
+        return "salt id:"+self._name+" mw:"+str(self._molew)+" melt:"+str(self._meltT)+" boil:"+str(self._boilT)+"\n    "+str(self._density)
+
+class TheInput:
+    Definition = None
+    @staticmethod
+    def definition():
+        if TheInput.Definition is not None: return TheInput.Definition
+        db = InputObject()
+        salts = db.createRequiredSingle("salts", Desc="The collection of salts in the system")
+        salts.addRequired("salt", Salt.definition())
+        salts.addUniqueConstraint(["salt/id"])
+        db.createRequiredSingle("queries", Desc="Parameters for queries salt properties") \
+            .createRequiredSingle("temperatures", Desc="Temperatures (C) at which to query density") \
+                .createRequired("value", MinValExc=0, Action=storeFloat)
+        TheInput.Definition = db
+        return TheInput.Definition
+
+    def createFrom(do:'DeserializedObject'):
+
+        result = TheInput()
+        result.salts = [Salt.createFrom(salt) for salt in do["salts"]["salt"]]
+        result.queryTemps = do["queries"]["temperatures"].valuelist()
+
+        return result
+
+    def __init__(self):
+        self.salts = None
+        self.queryTemps = None
+
+    def __str__(self):
+        return "".join(str(x)+"\n" for x in self.salts) + "\nqueried at: "+(",".join(str(t) for t in self.queryTemps))
 
 class TestSwigInterface(unittest.TestCase):
 
@@ -146,7 +254,7 @@ class TestSwigInterface(unittest.TestCase):
                 j += 1
             i += 1
 
-        #We are going to get "response_name_4" from the line "column 3 4 delimiter '-' as 'response_name_4' 
+        #We are going to get "response_name_4" from the line "column 3 4 delimiter '-' as 'response_name_4'
         for it in root:
             if it.name() == "extract_from" and str(it) == "filename_1.out":
                 for jt in it:
@@ -163,7 +271,7 @@ class TestSwigInterface(unittest.TestCase):
 
         doc = Interpreter(Syntax.SON,path="Not_Real.son")
         self.assertFalse(doc) # documents are not valid
-        parse_diagnostics = doc.parse_diagnostics()
+        parse_diagnostics = doc.parseDiagnostics()
         self.assertEqual(["Not_Real.son:1.1 is either inaccessible or doesn't exist! Unable to read."], [str(x) for x in parse_diagnostics])
 
 
@@ -331,6 +439,16 @@ class TestSwigInterface(unittest.TestCase):
                     MinValInc = 0
                 } % end value
             } % end B
+            C{
+                MinOccurs = 1
+                MaxOccurs = 1
+                value{
+                    MinOccurs = 1
+                    MaxOccurs = 1
+                    ValType   = Real
+                    MinValInc = 0
+                } % end value
+            } % end C
 
             MaxTemp{
                 MinOccurs = 1
@@ -338,7 +456,7 @@ class TestSwigInterface(unittest.TestCase):
                 value{
                     MinOccurs = 1
                     MaxOccurs = 1
-                    ValType   = Real 
+                    ValType   = Real
                     MinValExc = "../../MinTemp/value"
                 } % end value
             } % end MaxTemp
@@ -362,7 +480,7 @@ class TestSwigInterface(unittest.TestCase):
                     InputDefault = "linear"
                     MinOccurs = 1
                     MaxOccurs = 1
-                    ValType   = String 
+                    ValType   = String
                     ValEnums  = [ "linear" ]
                 } % end value
             } % end Type
@@ -400,7 +518,7 @@ queries{
     MaxOccurs = 1
 
     temperatures{
-        Description = "Temperatures (C) at which to query density" 
+        Description = "Temperatures (C) at which to query density"
         MinOccurs = 1
         MaxOccurs = 1
         value{
@@ -449,7 +567,7 @@ queries{
         self.assertEqual(salt[0].density(1200.0),1.77)
         self.assertEqual(salt[1].density(1200.0),1.9968)
 
-        # Test __getattr__ (. operator) through VectorWaspNode 
+        # Test __getattr__ (. operator) through VectorWaspNode
         # Each value belongs to different 'BoilTemp' requiring a VectorWaspNode
         boiltemps = doc.root().salts.salt.BoilTemp.value
         self.assertEqual(2, len(boiltemps))
@@ -459,7 +577,7 @@ queries{
         self.assertEqual("NaF", str(boiltemps[1].parent().parent().id))
 
     def test_validation_errors(self):
-        schema = '''data{MaxOccurs=1 value{MaxOccurs=1 MaxValInc=1.0 ValType=Real}}''' 
+        schema = '''data{MaxOccurs=1 value{MaxOccurs=1 MaxValInc=1.0 ValType=Real}}'''
         input = '''data 1.1'''
 
         # Test EDDI
@@ -501,7 +619,7 @@ queries{
         # Test HIT parse error
         interpreter = Interpreter(Syntax.HIT, data=input, path="myfile.inp")
         self.assertFalse(interpreter)
-        errors = interpreter.parse_diagnostics()
+        errors = interpreter.parseDiagnostics()
         expected = [
             "myfile.inp:1.7: syntax error, unexpected end of file",
             "myfile.inp:1.6: syntax error, 'data' has a missing or malformed value",
@@ -509,5 +627,165 @@ queries{
 
         error_strings = [str(x).rstrip() for x in errors]
         self.assertEqual(expected, error_strings)
+
+    def test_diagnostics(self):
+        input = "key=1.8"
+        interpreter = Interpreter(Syntax.SON, data=input, path="myfile.inp")
+        self.assertTrue(interpreter)
+        node = interpreter.root().key[0]['value'][0]
+        interpreter.createErrorDiagnostic(node, "is an unexpected parameter!")
+        interpreter.createWarningDiagnostic(node, "is an unexpected parameter!")
+        interpreter.createInformationDiagnostic(node, "is skipped as an unexpected parameter!")
+        interpreter.createHintDiagnostic(node, "should be named less ambiguously!")
+        dd = interpreter.deserializeDiagnostics()
+        self.assertEqual(4, len(dd))
+
+        expected = ["myfile.inp:1.5: value is an unexpected parameter!",
+                    "myfile.inp:1.5: value is an unexpected parameter!",
+                    "myfile.inp:1.5: value is skipped as an unexpected parameter!",
+                    "myfile.inp:1.5: value should be named less ambiguously!"]
+        for i, e in enumerate(expected):
+            self.assertEqual(e, str(dd[i]))
+
+    def test_database_inputobject(self):
+        '''This test exercises the input object definition creation, deserialization, selection, and diagnostics
+        as used by the toy salts classes above
+        '''
+        ss = '''
+        salts {
+          salt(LiF) {
+              MeltTemp : 1121.2
+              MolecularWeight : 25.9394
+              BoilTemp : 0
+              Density
+              {
+                  Type : foo
+                  A : 2.37
+                  B : 5.0e-4
+                  MinTemp : 1123.6
+                  MaxTemp : 1367.5
+              }
+          }
+          salt(NaF) {
+              MolecularWeight : 41.9882
+              MeltTemp : 1268
+              BoilTemp : 1978
+              Density
+              {
+                  Type : linear
+                  A : 2.76
+                  B : 6.36e-4
+                  MinTemp : 1273
+                  MaxTemp : 1373
+              }
+          }
+          salt(NaF){
+              MolecularWeight : 41.9882
+              MeltTemp : 1268
+              BoilTemp : 1978
+              Density
+              {
+                  Type : linear
+                  A : 2.76
+                  B : 6.36e-4
+                  MinTemp : 1273
+                  MaxTemp : 1373
+              }
+          }
+        }
+        queries {
+          temperatures = [1100,-1200,1300,1400]
+        }'''
+
+        interpreter = Interpreter(Syntax.SON, path="input.son", data=ss)
+
+        self.assertTrue(interpreter)
+
+        definition = TheInput.definition()
+        # Test input object (definition) selection logic
+        self.assertEqual(definition['salts']['salt']['MolecularWeight']['value'], definition.select("salts/salt/MolecularWeight/value")[0])
+        self.assertEqual(definition['salts']['salt']['MolecularWeight']['value'], definition.select("*/*/MolecularWeight/value")[0])
+        self.assertEqual(definition['salts']['salt']['MolecularWeight']['value'], definition.select("*/*/MolecularWeight/*")[0])
+        self.assertTrue(definition)
+
+        db = definition.deserialize(interpreter.root(), interpreter)
+        self.assertTrue(interpreter.deserializeDiagnostics())
+        expectedDiagnostics = '''input.son:6.26: value 0.0 is less than or equal to the allowed minimum exclusive value of 0!
+input.son:9.26: value foo is not one of the allows values ['linear']!
+input.son:16.16: id NaF must be unique but is duplicate to id on line 29 column 16
+input.son:29.16: id NaF must be unique but is duplicate to id on line 16 column 16
+input.son:44.32: value -1200.0 is less than or equal to the allowed minimum exclusive value of 0!
+'''
+        self.maxDiff = None
+        self.assertEqual(expectedDiagnostics,"".join(str(x)+"\n" for x in interpreter.deserializeDiagnostics()))
+
+        # Test deserialized result selection
+        saltids = db.select("salts/salt/id")
+        self.assertEqual(3, len(saltids))
+        self.assertEqual("LiF", saltids[0].storedResult())
+        self.assertEqual("NaF", saltids[1].storedResult())
+        self.assertEqual("NaF", saltids[2].storedResult())
+
+        self.assertTrue(db["salts"])
+        self.assertTrue(type(db["salts"]["salt"]) == list)
+        self.assertEqual(3, len(db["salts"]["salt"]))
+        self.assertTrue(db["salts"]["salt"][1])
+        self.assertTrue(db["salts"]["salt"][1]["MolecularWeight"])
+        self.assertEqual(41.9882, db["salts"]["salt"][1]["MolecularWeight"].value())
+
+        # Test defaulted 'C' and given A density coefficient
+        self.assertTrue(db["salts"]["salt"][1]["Density"])
+        self.assertTrue(db["salts"]["salt"][1]["Density"]["A"])
+        self.assertEqual(2.76, db["salts"]["salt"][1]["Density"]["A"].value())
+        self.assertTrue(db["salts"]["salt"][1]["Density"]["C"])
+        self.assertEqual(1.0, db["salts"]["salt"][1]["Density"]["C"].value())
+
+        self.assertTrue(db["queries"])
+        self.assertEqual([1100.0,-1200.0,1300.0,1400.0], db["queries"]["temperatures"].valuelist())
+
+        # create program data structure
+        theInput = TheInput.createFrom(db)
+        # This test illustrates the additional diagnostics from result to object creation
+        expectedDiagnostics = '''input.son:6.26: value 0.0 is less than or equal to the allowed minimum exclusive value of 0!
+input.son:9.26: value foo is not one of the allows values ['linear']!
+input.son:16.16: id NaF must be unique but is duplicate to id on line 29 column 16
+input.son:29.16: id NaF must be unique but is duplicate to id on line 16 column 16
+input.son:44.32: value -1200.0 is less than or equal to the allowed minimum exclusive value of 0!
+input.son:9.19: Type has value of foo which is not listed in ['linear']
+'''
+        self.assertEqual(expectedDiagnostics,"".join(str(x)+"\n" for x in interpreter.deserializeDiagnostics()))
+        self.assertEqual([1100.0,-1200.0,1300.0,1400.0],theInput.queryTemps)
+        self.assertEqual(3, len(theInput.salts))
+
+        # Test theInput for deserialized result
+        expectedSummary = '''salt id:LiF mw:25.9394 melt:1121.2 boil:0.0
+    density - a:2.37 b:0.0005 c:1.0 min:1123.6 max:1367.5
+salt id:NaF mw:41.9882 melt:1268.0 boil:1978.0
+    density - a:2.76 b:0.000636 c:1.0 min:1273.0 max:1373.0
+salt id:NaF mw:41.9882 melt:1268.0 boil:1978.0
+    density - a:2.76 b:0.000636 c:1.0 min:1273.0 max:1373.0
+
+queried at: 1100.0,-1200.0,1300.0,1400.0'''
+        self.assertEqual(expectedSummary, str(theInput))
+
+    def test_database_inputobject_exists(self):
+        definition = InputObject()
+        definition.createRequiredSingle("ids", Desc="List of viable identifers").createRequired("value", Action=storeStr)
+        definition.createRequiredSingle("ref", Desc="The id reference").createRequiredSingle("value", Action=storeStr)
+        definition.addExistsConstraint(ExistsConstraint(["ref/value"], target=["ids/value"]))
+
+        inputcontents = '''ids=[foo bar ted]
+        ref=bart
+        '''
+        interpreter = Interpreter(Syntax.SON, path="input.son", data=inputcontents)
+
+        self.assertTrue(interpreter)
+
+        db = definition.deserialize(interpreter.root(), interpreter)
+        self.assertTrue(interpreter.deserializeDiagnostics())
+        expectedDiagnostics = '''input.son:2.13: value bart is not one of the required existing targets! Required existing targets are foo, bar, ted!
+'''
+        self.maxDiff = None
+        self.assertEqual(expectedDiagnostics,"".join(str(x)+"\n" for x in interpreter.deserializeDiagnostics()))
 if __name__ == '__main__':
      unittest.main()